Exercises les 3 werkboek

+++ Exercises les 2/Exercise 2_4.py
@@ -1,3 +1,3 @@
 uurloon = float (input('wat verdien je per uur: '))
-uren = int (input('Hoeveel uur heb je gewerkt: '))
+uren = float (input('Hoeveel uur heb je gewerkt: '))
 print(uren, 'uur werken levert', uurloon*uren, 'euro op')
